+++ src/pages/Main.tsx
@@ -1,15 +1,21 @@
-import React, { useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import CardList from '../components/cardList/CardList';
 import ErrorBoundary from '../components/errorBoundary/ErrorBoundary';
 import ErrorMessage from '../components/errorBoundary/ErrorMessage';
 import Search from '../components/search/Search';
 import { useSearchQuery } from '../hooks/useSearchQuery';
 import Pagination from '../components/pagination/Pagination';
+import { getAllPeople, searchPeople } from '../services/api';
+import Loading from '../components/loading/Loading';
+import { Person } from '../utils/interfaces';
 
 const MainPage: React.FC = () => {
   const [query, setQuery] = useSearchQuery();
   const [hasError, setHasError] = useState<boolean>(false);
   const [page, setPage] = useState<number>(1);
+  const [totalCount, setTotalCount] = useState<number>(0);
+  const [people, setPeople] = useState<Person[]>([]);
+  const [isLoading, setIsLoading] = useState<boolean>(false);
 
   const handleSearch = (query: string) => {
     setQuery(query);
@@ -33,6 +39,28 @@ const MainPage: React.FC = () => {
     setPage(newPage);
   };
 
+  const loadFilms = async (query: string, page: number) => {
+    setIsLoading(true);
+    let data;
+    try {
+      if (query) {
+        data = await searchPeople(query, page);
+      } else {
+        data = await getAllPeople(page);
+      }
+      setPeople(data.results);
+      setTotalCount(data.count);
+    } catch (error) {
+      console.error('Error loading people:', error);
+    } finally {
+      setIsLoading(false);
+    }
+  };
+
+  useEffect(() => {
+    loadFilms(query, page);
+  }, [query, page]);
+
   if (hasError) {
     return <ErrorMessage onClose={closeErrorMessage} />;
   }
@@ -40,11 +68,18 @@ const MainPage: React.FC = () => {
   return (
     <ErrorBoundary onError={handleError}>
       <Search onSearch={handleSearch} onError={handleError} />
-      <Pagination
-        currentPage={page}
-        onPageChange={handlePageChange}
-      ></Pagination>
-      <CardList query={query} page={page} />
+      {isLoading ? (
+        <Loading />
+      ) : (
+        <>
+          <Pagination
+            currentPage={page}
+            onPageChange={handlePageChange}
+            hasMore={totalCount > page * 10}
+          />
+          <CardList people={people} />
+        </>
+      )}
     </ErrorBoundary>
   );
 };
